21.02.08_17:48
1.Hybrid_JSP_Connect : I just tried to connect the file that mad

2.PageControl _Images

3.PageControl_Number_Change_Color

4.LaunchScreen

5. PageControl_MapView: This map is that the version for Apple. So, you can see GPS and the point pin.

6. SwipeGesture_Images :
When I drag from right to left, the picture will change the next picture

7. PinchGesture : When I push option key, I can expand the size of the picture.

8.SegmentControl_MapView : There are three segment cotrol buttons. The first button is for the current location(GPS)
The other things are the usage method of the pin point

9. Swipe_Gesture :
There are four arrows and direction symbols. I just tried to make two functions when I touched the screen.
First of all, when I use one finger, I can drag in cardinal directions and it will be changed the black color to red color.
The other thing that I made is after I push option key plus shift key and drag any direction, the color will be changed the black to green